Import key signatures from Guitar Pro files.
This also corrects a couple bugs in the original implementation:
- Cancellations are now handled.
- Key signatures with flats are now handled.

+++ source/formats/guitar_pro/document.h
@@ -157,7 +157,7 @@ struct Measure
     boost::optional<int> myRepeatEnd;
     /// If there is a key change, this contains the key and whether it is
     /// minor.
-    boost::optional<std::pair<int, bool>> myKeyChange;
+    boost::optional<std::pair<int8_t, bool>> myKeyChange;
     /// Optional rehearsal sign.
     boost::optional<std::string> myMarker;
     /// Optional alternate ending number.
